Firmware/Software: Implemented sparse memory model for screen contents
Instead of allocating a complete character buffer for the whole screen now
memory is only allocated for those portions of the screen that are actually
used. The amount of memory is determined from the write/clear commands
defined in the configuration.

In addition, the parser now calculates and reports the amount of memory
allocated for a given configuration on the host.