Firmware: Added global output enable pin, improved control logic
The former OE (output enable) pin has been renamed to OR (output request)
and another pin has been introduced as OE (output enable).

The output enable pin now directly controls whether or not the display is
active. When low, the display is enabled. When going high again, the
display is immediately disabled, regardless of any running timeouts from a
former output request.

This allows using the OE line like a switch to turn the display on and off
while a low pulse on OR can be used to temporarily show the current state
for a certain time.