Software/Firmware: Added "font" command to flash alternative font
The firmware needs to be rebuild and updated for this to work.

overlay64 font <infile>

The input file is a binary font image similar to a C64 rom image.
Only the first 96 characters are flashed, corresponding to the
ascii characters 32-127.

To convert an existing C64 font (which has a different order),
use the fontconv bash script:

./fontconv.sh charset.rom overlay.rom