Software: Copy only common chars from c64 to overlay64 font
When converting from a C64 charset to an overlay64 font, only
the glyphs common in both ascii and petscii are copied from the
C64 charset. The glyphs for backspace, circumflex, underscore,
backtick, curly braces, pipe and tilde are copied from the
default overlay font instead. This way the resulting font is a
proper ascii font for the ascii codes 32-127.