Software/Firmware: Improved display logic, improved flashing, bugfixes
Improved screen logic. Independent of the screen or control mode,
a screen is only enabled if at least one write command is found
to be effective. This can either be an unconditional write command
or a write command depending on a certain input line state.
In addition, a sample does not need to cover all possible input
line states anymore.

The new screen mode "always" has been added to allow a screen to
be always enabled regardless of control lines being asserted or
input lines being changed, as far as at least one write command
is effective.

Combining both features allows for text to be displayed
permanently when one or more input lines have a specific
state.

For example, the following setup would display text only as long
as the reset line is low:

RESET=0

screen always
  sample RESET
    when 0 write 0 14 "Reset..."

The configuration is now flashed to eeprom directly from the
application, so that entering the bootloader beforehand is no
longer necessary. Flashing is now also faster for small changes
and easier on the eeprom since only changed values are updated.

The exit codes of the command line utility have been corrected.
The utility now reports success or failure correctly.