Firmware: Improved chances of proper device enumeration for some Windows versions.
It appears that on some Windows versions and/or installations device
enumeration may fail due to Windows giving up on enumeration too early.

A workaround for the current 1.1 version of the firmware is to disconnect
the cable from the PC when the device is powered on and to connect
it to the PC after a few seconds.

Thus the previously introduced delay has been removed again. Note
that enumeration on Linux is/was not negatively affected by either
variant.

Thanks Richard van der Wolf for pointing out and also helping to
resolve this issue.