Firmware: Fixed slightly skewed first text lines on some monitors
On some monitors, the display of the first text line may have been
slightly skewed due to the fact that the SPI line was simply left
enabled and low throughout the next hsync. Whether or not this
effect is noticiable seems to depend on the actual input impedance
of the monitor in use. The SPI line is now tristated after every
rendered scanline, shortly before the next hsync occurs.