Allow unconditional writes within samples with interpolation of values
It is now possible to to add an unconditional write command within the
scope of a sample (i.e. before any when-clause) that will be executed
regardless of the sampled value if the containing screen is enabled.

The sampled value can be interpolated into the string written by this
write command by using up to 16 printf-style format specifications.

For example, if the user wishes to display the value of four input lines
as a decimal value the following configuration can be used:

screen
  sample 0 1 2 3
    write 0 0 "THE VALUE IS %d"

Previously this was only possible by handling every single value manually:

screen
  sample 0 1 2 3
    when 0000 write "THE VALUE IS 0"
    when 0001 write "THE VALUE IS 1"
    when 0010 write "THE VALUE IS 2"
    [...]

PLEASE NOTE THAT THIS COMMIT BREAKS BINARY COMPATIBILITY WITH EXISTING
CONFIGURATION STORED ON THE DEVICE.

This means that before updating to this firmware version, the EEPROM of the
microcontroller must be erased and the configuration needs to be recreated
and reflashed using a newly build client application.

The update command will allow a more convenient method of updating the
firmware in case of binary incompatibilities in the next official release.