Improved update command for updates that break binary compatibility
The update command now accepts a configuration file as a second optional
argument. If specified, the exisitng configuration on the device will be
erased before updating the firmware and recreated and reflashed after the
update. This prevents the new firmware from crashing when trying to read
a deprecated binary configuration format still present on the device.